pull some methods out of the manager.go file
