+++ pkg/manager/manager.go
@@ -400,108 +400,6 @@ func (m MediaManager) DeleteIndexer(ctx context.Context, request DeleteIndexerRe
 	return m.storage.DeleteIndexer(ctx, int64(*request.ID))
 }
 
-type AddQualityDefinitionRequest struct {
-	model.QualityDefinition
-}
-
-// AddQualityDefinition stores a new quality definition in the database
-func (m MediaManager) AddQualityDefinition(ctx context.Context, request AddQualityDefinitionRequest) (model.QualityDefinition, error) {
-	definition := request.QualityDefinition
-
-	id, err := m.storage.CreateQualityDefinition(ctx, request.QualityDefinition)
-	if err != nil {
-		return definition, err
-	}
-
-	definition.ID = int32(id)
-	return definition, nil
-}
-
-// DeleteQualityDefinitionRequest request to delete a quality definition
-type DeleteQualityDefinitionRequest struct {
-	ID *int `json:"id" yaml:"id"`
-}
-
-// AddQualityDefinition deletes a quality definition
-func (m MediaManager) DeleteQualityDefinition(ctx context.Context, request DeleteQualityDefinitionRequest) error {
-	if request.ID == nil {
-		return fmt.Errorf("indexer id is required")
-	}
-
-	return m.storage.DeleteQualityDefinition(ctx, int64(*request.ID))
-}
-
-// ListQualityDefinitions list stored quality definitions
-func (m MediaManager) ListQualityDefinitions(ctx context.Context) ([]*model.QualityDefinition, error) {
-	return m.storage.ListQualityDefinitions(ctx)
-}
-
-// ListQualityDefinitions get a stored quality definitions
-func (m MediaManager) GetQualityDefinition(ctx context.Context, id int64) (model.QualityDefinition, error) {
-	return m.storage.GetQualityDefinition(ctx, id)
-}
-
-func (m MediaManager) GetQualityProfile(ctx context.Context, id int64) (storage.QualityProfile, error) {
-	return m.storage.GetQualityProfile(ctx, id)
-}
-
-func (m MediaManager) ListQualityProfiles(ctx context.Context) ([]*storage.QualityProfile, error) {
-	return m.storage.ListQualityProfiles(ctx)
-}
-
-// AddDownloadClientRequest describes what is required to add a download client
-type AddDownloadClientRequest struct {
-	model.DownloadClient
-}
-
-func (m MediaManager) CreateDownloadClient(ctx context.Context, request AddDownloadClientRequest) (model.DownloadClient, error) {
-	downloadClient := request.DownloadClient
-
-	id, err := m.storage.CreateDownloadClient(ctx, request.DownloadClient)
-	if err != nil {
-		return downloadClient, err
-	}
-
-	downloadClient.ID = int32(id)
-	return downloadClient, nil
-}
-
-func (m MediaManager) GetDownloadClient(ctx context.Context, id int64) (download.DownloadClient, error) {
-	client, err := m.storage.GetDownloadClient(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	return m.factory.NewDownloadClient(client)
-}
-
-func (m MediaManager) ListDownloadClients(ctx context.Context) ([]*model.DownloadClient, error) {
-	return m.storage.ListDownloadClients(ctx)
-}
-
-func (m MediaManager) DeleteDownloadClient(ctx context.Context, id int64) error {
-	return m.storage.DeleteDownloadClient(ctx, id)
-}
-
-func availableProtocols(clients []*model.DownloadClient) map[string]struct{} {
-	ret := make(map[string]struct{})
-	for _, c := range clients {
-		ret[c.Type] = struct{}{}
-	}
-
-	return ret
-}
-
-func clientForProtocol(clients []*model.DownloadClient, proto prowlarr.DownloadProtocol) *model.DownloadClient {
-	for _, c := range clients {
-		if c.Type == string(proto) {
-			return c
-		}
-	}
-
-	return nil
-}
-
 func nullableDefault[T any](n nullable.Nullable[T]) T {
 	var def T
 	if n.IsSpecified() {
